% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cuml_ovpaymt_table}
\alias{cuml_ovpaymt_table}
\title{Cumulative percent overpayment in sample as a function of score
Ordered by score, descending}
\usage{
cuml_ovpaymt_table(df)
}
\description{
Cumulative percent overpayment in sample as a function of score
Ordered by score, descending
}

