% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cuml_response_table}
\alias{cuml_response_table}
\title{Cumulate % captured response vs. culuative % claimants
Ordered by score, descending}
\usage{
cuml_response_table(df)
}
\description{
Cumulate % captured response vs. culuative % claimants
Ordered by score, descending
}

