% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn_validate}
\alias{knn_validate}
\title{Run KNN validation script and push tables and plots to global environment}
\usage{
knn_validate(test.df)
}
\arguments{
\item{test.df}{Overpayment-enriched data frame for testing.
See the \code{APM} documentation for, e.g., \code{knn_train()} and
\code{make_sample_table_with_n_records()}.}
}
\description{
Run KNN validation script and push tables and plots to global environment
}

