% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{moving_avg_table}
\alias{moving_avg_table}
\title{Percent overpayment moving average vs. cumulative percent claimants
Ordered in the same way as cumulative percent response}
\usage{
moving_avg_table(df, n.pts = 1001)
}
\description{
Percent overpayment moving average vs. cumulative percent claimants
Ordered in the same way as cumulative percent response
}

