% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cml.pct.ov_per_score}
\alias{plot_cml.pct.ov_per_score}
\title{Plot the cumulative % overpayments vas. score (descending)
I.e., where should our cutoff in score be placed to capture an ideal
amount of overpayment?}
\usage{
plot_cml.pct.ov_per_score(df, model = "")
}
\description{
Plot the cumulative % overpayments vas. score (descending)
I.e., where should our cutoff in score be placed to capture an ideal
amount of overpayment?
}

