% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cml.res_per_cml.clmt}
\alias{plot_cml.res_per_cml.clmt}
\title{Plot the cumulative % response vs. cumulative % claimants}
\usage{
plot_cml.res_per_cml.clmt(df, model = "")
}
\description{
Plot the cumulative % response vs. cumulative % claimants
}

