% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_moving_avg}
\alias{plot_moving_avg}
\title{Plot the moving average of % overpayment per score}
\usage{
plot_moving_avg(df, model = "")
}
\description{
Plot the moving average of % overpayment per score
}

