% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pct.ov_per_score}
\alias{plot_pct.ov_per_score}
\title{Plot the % overpayment per score
This is not terribly useful because we may have very few observations
at very high scores. On the other hand, if we have 0 overpayments at
high scores, that might mean something.}
\usage{
plot_pct.ov_per_score(df, model = "")
}
\description{
Plot the % overpayment per score
This is not terribly useful because we may have very few observations
at very high scores. On the other hand, if we have 0 overpayments at
high scores, that might mean something.
}

