% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pct.ov_per_score_bin}
\alias{plot_pct.ov_per_score_bin}
\title{Plot the percent overpayment per score bin
Binned into units of 10 points; again, not especially useful, but could
indicate problems with models if we do very poorly at high scores}
\usage{
plot_pct.ov_per_score_bin(df, model = "")
}
\description{
Plot the percent overpayment per score bin
Binned into units of 10 points; again, not especially useful, but could
indicate problems with models if we do very poorly at high scores
}

