% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf_validate}
\alias{rf_validate}
\title{Run RF validation script and push tables and plots to global environment}
\usage{
rf_validate(test.df)
}
\arguments{
\item{test.df}{Overpayment-enriched data frame for testing.
See the \code{APM} documentation for, e.g., \code{knn_train()} and
\code{make_sample_table_with_n_records()}.}
}
\description{
Run RF validation script and push tables and plots to global environment
}

