% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{scores_table}
\alias{scores_table}
\title{Percent overpayment per score
Probably only useful for "discrete" scores}
\usage{
scores_table(df)
}
\description{
Percent overpayment per score
Probably only useful for "discrete" scores
}

