% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{sql_module_df}
\alias{sql_module_df}
\title{Pull scores for appropriate module from nrdb then
merge with target from aggr_claimt_ovpaymt}
\usage{
sql_module_df(sql_module, cycle_date = NULL, prevDays = 365,
  onlyFraud = FALSE)
}
\arguments{
\item{sql_module}{The module you want to pull data for}

\item{cycle_date}{Pretty obvious at this point}

\item{prevDays}{Number of days back the overpayment could have been occuring}

\item{onlyFraud}{if True then we only count Fraud as a 1, not just overpayment}
}
\value{
Data frame containing all the scores for all the claims from nrd.enty_score

for the desired module along with the target value.
}
\description{
Pull scores for appropriate module from nrdb then
merge with target from aggr_claimt_ovpaymt
}

