% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_validate}
\alias{sql_validate}
\title{Run SQL validation script and push tables and plots to global environment}
\usage{
sql_validate(score_cd = c("NWC_COR", "CLA_RISK", "QWR_QWR", "IPA_CHIP",
  "IPA_SHIP_PERD", "IPA_SHIP_RAW", "IPA_SHIP_ALL", "IPA_ELR", "IPA_SHIP",
  "IPA_SHIP_TIME", "IPA_ELR_PCT", "IPA_ELR_TOT", "IPA_CLR", "NHI_NHR",
  "SOM_SOM", "SVM_SVM", "RF_RF", "KNN_KNN"), prevDays = 365,
  onlyFraud = FALSE)
}
\description{
By default will use the scores from the most recent cycle date to predict
whether claimants had overpayment over the previous 365 days. Can be changed
to any length of time prior to the cycle date and also whether or not only
fraud should be considered (instead of overpayment).
}

